requirements.txt
flask
gunicorn
kiteconnect
python-dotenv
requests
numpy
pytz
